fix: split cli openssl and php openssl tests to two files
